Part 2: Timestamp 4
2:49:00 - User registration

+++ views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -14,8 +16,10 @@ from .forms import RoomForm
  #   {'id': 3, 'name': 'Frontend Developers'},
 #]
 
-
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -34,15 +38,12 @@ def loginPage(request):
         else:
             messages.error(request, 'Username OR password does not exit')
 
-
     context = {}
     return render(request, 'base/login_register.html', context)
-
 
 def loginUser(request):
     logout(request)
     return redirect('home')
-
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -59,8 +60,6 @@ def home(request):
     context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, 'base/home.html', context)
 
-
-
 def room(request, pk):
     room = Room.objects.get(id=pk)
 
@@ -68,8 +67,10 @@ def room(request, pk):
 
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
+
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -79,9 +80,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -92,9 +97,15 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
-    return render(request, 'base/delete.html', {'odj':room})
+        
+    return render(request, 'base/delete.html', {'obj': room})
